python/requirements.txt
flask
azure-identity==1.7.0
pyraisdk==0.1.0
GPUtil
